# Improve performance of `delegate_hashed_bins` when delegating to a large number of bins