**Fixes issue #**: Quick fix for #1005 

**Description of the changes being introduced by the pull request**:
* Factor out `delegate()` logic into various helpers:
  * `_keys_to_keydict()`
  * `_create_delegated_target()`
  * `_update_roledb_delegations()`
* Use those helpers directly in `delegated_hashed_bins()`, rather than calls to `delegate()`, most significantly queuing up changes to roledb and performing them in one operation with `_update_roledb_delegations()`

Based on the profiling done by @lukpueh after @woodruffw's report we gain a significant speed improvement when delegating to a large number of hashed bins (i.e. 16k):
before: `python3 profile_tuf_bins.py  4978.27s user 39.37s system 99% cpu 1:24:00.47 total`
after: `python3 profile_tuf_bins.py  21.09s user 10.61s system 96% cpu 32.979 total`


**Please verify and check that the pull request fulfils the following
requirements**:

- [ ] The code follows the [Code Style Guidelines](https://github.com/secure-systems-lab/code-style-guidelines#code-style-guidelines)
- [ ] Tests have been added for the bug fix or new feature
- [ ] Docs have been added for the bug fix or new feature