@joshuagl 2020-04-03
> Thanks for the changes, @joshuagl. The performance gain is uncanny! :)

Thanks largely to your profiling 🙂 

> I have a few minor suggestions nothing blocking though. Let me know if you want to change anything, otherwise we can merge this.

All great suggestions, I've squashed them into the original patches. Thanks for the thoughtful review @lukpueh